consensus: refactor tests
